move CharcoalForegroundColor into Extensions folder
