+++ Sources/Charcoal/SwiftUI/Colors/OnSurfaces.swift
@@ -1,21 +1,6 @@
 import SwiftUI
 
 @available(iOS 13, *)
-struct CharcoalForegroundColor: ViewModifier {
-    let color: Color
-    
-    func body(content: Content) -> some View {
-        if #available(iOS 15, *) {
-            content
-                .foregroundStyle(color)
-        } else {
-            content
-                .foregroundColor(color)
-        }
-    }
-}
-
-@available(iOS 13, *)
 struct CharcoalOnSurfaceBorder: ViewModifier {
     func body(content: Content) -> some View {
         content
@@ -143,9 +128,3 @@ public extension View {
     }
 }
 
-@available(iOS 13, *)
-public extension View {
-    func charcoalForegroundColor(_ color: Color) -> some View {
-        modifier(CharcoalForegroundColor(color: color))
-    }
-}
